backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-groq
openmetadata-ingestion
python-dotenv
httpx
